requirements.txt
axial_positional_embedding==0.3.11
# backcall @ file:///home/conda/feedstock_root/build_artifacts/backcall_1592338393461/work
clarabel==0.9.0
CoLT5-attention==0.11.1
contourpy==1.1.1
cvxpy==1.5.2
cycler==0.12.1
# decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
ecos==2.0.14
einops==0.8.1
# entrypoints @ file:///home/conda/feedstock_root/build_artifacts/entrypoints_1643888246732/work
et_xmlfile==2.0.0
filelock==3.16.1
fonttools==4.56.0
fsspec==2025.2.0
importlib_resources==6.4.5
# ipykernel @ file:///home/conda/feedstock_root/build_artifacts/ipykernel_1620912944151/work/dist/ipykernel-5.5.5-py3-none-any.whl
# ipython @ file:///home/conda/feedstock_root/build_artifacts/ipython_1651240558244/work
# ipython_genutils @ file:///home/conda/feedstock_root/build_artifacts/ipython_genutils_1716278396992/work
# jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1669134318875/work
Jinja2==3.1.5
joblib==1.4.2
# jupyter-client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1633454794268/work
# jupyter_core @ file:///home/conda/feedstock_root/build_artifacts/jupyter_core_1727163409502/work
kiwisolver==1.4.7
local-attention==1.10.0
MarkupSafe==2.1.5
matplotlib==3.7.5
# matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1713250518406/work
mpmath==1.3.0
narwhals==1.26.0
# nest_asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1705850609492/work
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.8.61
nvidia-nvtx-cu12==12.1.105
openpyxl==3.1.5
osqp==0.6.7.post3
packaging==24.2
pandas==2.0.3
# parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1712320355065/work
patsy==1.0.1
# pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1706113125309/work
# pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
pillow==10.4.0
# platformdirs @ file:///home/conda/feedstock_root/build_artifacts/platformdirs_1726613481435/work
plotly==6.0.0
product_key_memory==0.2.11
# prompt_toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1717574682739/work
properscoring==0.1
# ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1609419310487/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
# Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1714846767233/work
pyparsing==3.1.4
# python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
pytz==2025.1
PyWavelets==1.4.1
pyzmq==19.0.2
qdldl==0.1.7.post5
reformer-pytorch==1.4.4
scikit-learn==1.3.2
scipy==1.10.1
scs==3.2.7.post2
# six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
sklearn-quantile==0.0.29
statsmodels==0.14.1
sympy==1.13.3
threadpoolctl==3.5.0
torch==2.4.1
torchaudio==2.4.1
torchvision==0.19.1
# tornado @ file:///home/conda/feedstock_root/build_artifacts/tornado_1648827257044/work
tqdm==4.67.1
# traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1713535121073/work
triton==3.0.0
typing_extensions==4.12.2
tzdata==2025.1
# wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1704731205417/work
xarray==2023.1.0
xgboost==2.1.4
xlrd==2.0.1
zipp==3.20.2
